refactor: rename DslConfig to avoid re-naming
