+++ octopus/dsl/__init__.py
@@ -2,6 +2,6 @@
 DSL module for test configuration and parsing
 """
 
-from . import config, parser, validator
+from . import dsl_config, parser, validator
 
-__all__ = ["config", "parser", "validator"]
+__all__ = ["dsl_config", "parser", "validator"]
